refactor: extract waitlist form into a separate component
